requirements.txt
blinker==1.8.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
gunicorn==22.0.0
PyJWT==2.9.0
MarkupSafe==2.1.5
pycparser==2.22
pymongo==4.8.0
Werkzeug==3.0.3
python-dotenv==1.0.1
